+++ src/main/java/com/kputimcevs/insurance/premium/entities/RiskType.java
@@ -4,15 +4,15 @@ import static com.kputimcevs.insurance.premium.entities.ThresholdOperator.ABOVE;
 import static com.kputimcevs.insurance.premium.entities.ThresholdOperator.ABOVE_OR_EQUAL;
 
 public enum RiskType {
-    FIRE(100, ABOVE, 0.014, 0.028),
-    THEFT(15, ABOVE_OR_EQUAL, 0.11, 0.05);
+    FIRE(100, ABOVE, (float) 0.014, (float) 0.028),
+    THEFT(15, ABOVE_OR_EQUAL, (float) 0.11, (float) 0.05);
 
-    public final double threshold;
-    public final double averageCoefficient;
-    public final double aboveThresholdCoefficient;
+    public final float threshold;
+    public final float averageCoefficient;
+    public final float aboveThresholdCoefficient;
     public final ThresholdOperator thresholdOperator;
 
-    RiskType(int threshold, ThresholdOperator thresholdOperator, double averageCoefficient, double aboveThresholdCoefficient) {
+    RiskType(int threshold, ThresholdOperator thresholdOperator, float averageCoefficient, float aboveThresholdCoefficient) {
         this.threshold = threshold;
         this.averageCoefficient = averageCoefficient;
         this.aboveThresholdCoefficient = aboveThresholdCoefficient;
